+++ src/main/java/com/bemba/goalsapi/dto/GoalDto.java
@@ -15,7 +15,7 @@ public class GoalDto {
 	
 	private LocalDate deadline;
 	
-	private Long totalHours;
+	private Double totalHours;
 	
 	private String rewardName;
 }
